Programa y correcciones de aritmetica float

+++ Objetos/TP2022_ProyectoAlumnos/src/tema4/Entrenador.java
@@ -25,7 +25,7 @@ public class Entrenador extends Empleado {
     
     @Override
     public double calcularEfectividad(){
-        return this.getCantidadTitulos() / this.getAntiguedad();
+        return (double)this.getCantidadTitulos() / (double)this.getAntiguedad();
     }
     @Override
     public double calcularSueldoACobrar(){
